pkg/testhelper: move pod utilities to a package
This avoids circular imports when `pkg/kubernetes` uses `pkg/testhelper` in its
tests.
